+++ rag_quest/worlds/__init__.py
@@ -2,6 +2,15 @@
 
 from .exporter import WorldExporter
 from .importer import WorldImporter
+from .modules import Module, ModuleManifestError, ModuleRegistry, load_modules
 from .templates import STARTER_WORLDS
 
-__all__ = ["WorldExporter", "WorldImporter", "STARTER_WORLDS"]
+__all__ = [
+    "WorldExporter",
+    "WorldImporter",
+    "STARTER_WORLDS",
+    "Module",
+    "ModuleRegistry",
+    "ModuleManifestError",
+    "load_modules",
+]